+++ noodles-bcf/CHANGELOG.md
@@ -7,6 +7,8 @@
   * bcf/writer/vcf_writer/genotypes: Add FORMAT GT encoder for missing alleles
     (i.e., `.`).
 
+  * bcf/writer/vcf_writer/genotypes: Pad integer vector for genotypes.
+
 ## 0.19.1 - 2022-11-29
 
 ### Fixed